fix: Resolve query variables when creating Fields
