+++ core/src/main/scala/caliban/validation/Validator.scala
@@ -101,6 +101,7 @@ object Validator {
                 op.selectionSet,
                 fragments,
                 variables,
+                op.variableDefinitions,
                 operation.opType,
                 document.sourceMapper,
                 op.directives,
@@ -530,6 +531,7 @@ object Validator {
                     op.selectionSet,
                     context.fragments,
                     Map.empty[String, InputValue],
+                    List.empty[VariableDefinition],
                     t,
                     SourceMapper.empty,
                     Nil,
